# Refactor LST inputs