--- a/HLTrigger/Configuration/python/HLT_75e33/modules/hltPhase2OTHitsInputLST_cfi.py
+++ b/HLTrigger/Configuration/python/HLT_75e33/modules/hltPhase2OTHitsInputLST_cfi.py
@@ -1,6 +1,0 @@
-import FWCore.ParameterSet.Config as cms
-
-hltPhase2OTHitsInputLST = cms.EDProducer('LSTPhase2OTHitsInputProducer',
-    phase2OTRecHits = cms.InputTag('hltSiPhase2RecHits'),
-    mightGet = cms.optional.untracked.vstring
-)