--- a/RecoTracker/LST/python/lst_cff.py
+++ b/RecoTracker/LST/python/lst_cff.py
@@ -1,6 +1,5 @@
 import FWCore.ParameterSet.Config as cms
 
 from RecoTracker.LST.lstSeedTracks_cff import *
-from RecoTracker.LST.lstPixelSeedInputProducer_cfi import *
-from RecoTracker.LST.lstPhase2OTHitsInputProducer_cfi import *
+from RecoTracker.LST.lstInputProducer_cfi import *
 from RecoTracker.LST.lstOutputConverter_cfi import *