--- a/RecoTracker/LST/src/classes.h
+++ b/RecoTracker/LST/src/classes.h
@@ -2,8 +2,6 @@
 #define RecoTracker_LST_classes_h
 
 #include "DataFormats/Common/interface/Wrapper.h"
-#include "RecoTracker/LST/interface/LSTPixelSeedInput.h"
-#include "RecoTracker/LST/interface/LSTPhase2OTHitsInput.h"
 #include "RecoTracker/LST/interface/LSTOutput.h"
 
 #endif