This PR improves a bit the way inputs are passed into LST. Portable collections are constructed from the very beginning, and the code used in CMSSW and standalone was unified. There are some points to discuss that I'll point out in review comments.