@VourMa 2025-03-17
> Standalone works now, but I think I'll need some help from @VourMa to get CMSSW to work. I'm currently getting an error saying that there is no dictionary for the new LSTInput collection. I'm not sure what it is, but my guess is that it's probably something with the fillDescriptions method of LSTInputProducer.

I can take a look later today/tomorrow. I will message you for the details.