@ariostas 2025-03-17
In the meantime, here's a timing comparison on cgpu-1:

```
This PR [0bb4f24]
Average time for map loading = 582.717 ms
Average time for input loading = 7165.74 ms
Average time for lst::Event creation = 0.000442778 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      2.3      0.4      0.4      1.6      1.3      0.6      1.0      0.8      1.2      0.0       9.5       6.6+/-  1.4       9.6   explicit[s=1]
   avg      2.7      0.8      0.7      2.1      1.8      0.9      1.6      1.0      1.8      0.0      13.4       9.9+/-  2.2       6.8   explicit[s=2]
   avg      3.6      1.3      1.4      3.1      2.6      1.1      2.8      1.7      3.4      0.0      21.1      16.3+/-  3.2       5.4   explicit[s=4]
   avg      4.7      2.2      2.2      4.3      3.8      1.5      3.9      2.2      4.8      0.1      29.7      23.4+/-  4.7       5.1   explicit[s=6]
   avg      6.2      3.4      3.0      5.6      5.3      1.9      5.2      2.8      6.3      0.0      39.7      31.6+/-  6.9       5.1   explicit[s=8]

Master [38eca30]
Average time for map loading = 589.504 ms
Average time for input loading = 7417.81 ms
Average time for lst::Event creation = 0.000488191 ms
   Evt    Hits       MD       LS      T3       T5       pLS       pT5      pT3      TC       Reset    Event     Short             Rate
   avg      2.3      0.4      0.4      1.6      1.3      0.6      1.0      0.8      1.2      0.0       9.6       6.6+/-  1.4       9.6   explicit[s=1]
   avg      1.7      0.7      0.7      2.0      1.6      0.8      1.4      0.9      1.9      0.0      11.7       9.2+/-  1.9       5.9   explicit[s=2]
   avg      2.8      1.3      1.4      3.1      2.7      1.1      2.7      1.4      3.3      0.0      19.9      16.0+/-  3.2       5.1   explicit[s=4]
   avg      4.1      2.1      2.2      4.3      3.9      1.5      3.7      2.0      4.9      0.0      28.9      23.3+/-  4.0       4.9   explicit[s=6]
   avg      5.3      3.2      3.1      5.7      5.6      2.0      5.2      2.6      6.5      0.0      39.3      32.0+/-  6.2       5.0   explicit[s=8]
```

The slowdown on the hits column for multiple streams is probably from this part. Any suggestions for a way to fix it? The options I had in mind were just creating an extra vector with references to the right queue, or less preferably, making the queue a public attribute.
https://github.com/SegmentLinking/cmssw/blob/0bb4f24862a5cd6f76b923387166e5931d6e5bdb/RecoTracker/LSTCore/standalone/bin/lst.cc#L408-L413