@ariostas 2025-03-19
I added the class defs, but I'm still getting the same error. I'm not sure what I'm missing