@VourMa 2025-03-19
> I added the class defs, but I'm still getting the same error. I'm not sure what I'm missing

I noticed that you added the class definition in the `LST` package. Don't you define the class in the `LSTCore` package though?