@ariostas 2025-03-25
> I took a look at the code and made some comments, mostly for my understanding.
>
> Does the class issue remain in CMSSW? If so, let me know and I could make some time to run the code myself and try to figure out the problem.

Thank you for your comments, @VourMa! I replied to them. I'm still having issues with classes, but I'm giving it another try. In the meeting later today I'll let you know how it goes.