@ariostas 2025-03-25
I think standalone should be fine now.

/run standalone