@slava77 2025-03-28
maybe it's easier to submit this as a draft upstream and get feedback there?
I'm not sure the last exchange in https://mattermost.web.cern.ch/cms-o-and-c/pl/jo9f68nzwbnmmkh38dr3qtf3fa was productive