@ariostas 2025-03-28
> maybe it's easier to submit this as a draft upstream and get feedback there?

Sounds good, I'll rename things as you suggested and make sure that the rest is fine, and then I'll submit it. However, I'll be on Holiday until Tuesday, so I'll do so when I get back