@ariostas 2025-04-10
@VourMa I think this is ready for you to fix the HLT part.

We still have to decide on the naming and handling of the SoAs that were split. @slava77 should I rename `Input*`, `*` to `*Base`, `*Extended`? And should I define a small struct that stores pointers to these two?