@slava77 2025-04-10
maybe it's easier to submit this as a draft upstream and get feedback there?
I'm not sure the last exchange in https://mattermost.web.cern.ch/cms-o-and-c/pl/jo9f68nzwbnmmkh38dr3qtf3fa was productive


> should I rename `Input*`, `*` to `*Base`, `*Extended`? 

I didn't come up with a better suggestion since my earlier review.

> And should I define a small struct that stores pointers to these two?

I don't have a strong preference for this: without writing extra methods that would encapsulate this union, it will not be particularly practical; but to encapsulate may need too much coding.