@VourMa 2025-04-11
> @VourMa I think this is ready for you to fix the HLT part.

Thanks! Let's discuss in the meeting and I can prepare this for early next week (I would like to see also the bot results in CMSSW).