@ariostas 2025-04-15
On Manos's branch even standalone was segfaulting, so there must have been some aditional issue. I rebased to the current master and fixed an inconsistency. Standalone seems to be fine, so the issue is probably from the input producer. I'll run the CI to see if the CMSSW plots look any different.

/run all