@ariostas 2025-04-17
The error was because I needed to adjust the CI.

@VourMa the HLT setup that you had in the other PR works too. (Or at least it runs without crashing, but I haven't looked at the plots)