@VourMa 2025-04-17
> @VourMa the HLT setup that you had in the other PR works too. (Or at least it runs without crashing, but I haven't looked at the plots)

Thanks! @slava77 Should we rerun the tests on cmssw PR to check also for differences in results?