@VourMa 2025-04-19
> Thanks! @slava77 Should we rerun the tests on cmssw PR to check for differences in results?

It seems like 0.755 shows differences in the trigger results (even though I don't fully get the output). I think we need to run the HLT plots to make sure what's going on.